entry: SDK and Types updates (#284)
