+++ packages/sdk/src/index.ts
@@ -17,5 +17,5 @@ export * as Utils from '@cord.network/utils'
 // export * as Namespace from '@cord.network/namespace'
 export * as Profile from '@cord.network/profile'
 export * as Registry from '@cord.network/registry'
+export * as Entry from '@cord.network/entry'
 export * from '@cord.network/types'
-
