+++ packages/types/src/index.ts
@@ -29,3 +29,4 @@ export * from './Namespace.js'
 export * from './ChainSpace.js'
 export * from './Statement.js'
 export * from './Registry.js'
+export * from './Entry.js'
